+++ frontend/src/landing_page/landing_page.test.ts
@@ -1,0 +1,3 @@
+describe('landingPage', () => {
+  it.todo('works');
+});
