+++ frontend/src/loading_page/loading_page.test.ts
@@ -1,0 +1,3 @@
+describe('loadingPage', () => {
+  it.todo('works');
+});
