+++ frontend/src/main/main.test.ts
@@ -1,0 +1,3 @@
+describe('main', () => {
+  it.todo('works');
+});
